+++ src/main/java/https/github/com/wallas5h/LaskoMed/infrastructure/database/repository/jpa/ReferralJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReferralJpaRepository extends JpaRepository<ReferralEntity, Integer> {
+public interface ReferralJpaRepository extends JpaRepository<ReferralEntity, Long> {
 
 }
